+++ R/sienaprint.r
@@ -268,12 +268,9 @@ print.sienaFit <- function(x, tstat=TRUE, ...)
 		}
 		else
 		{
-			if (!is.null(x$gmm))
+			if (gmm(x))
 			{
-				if (x$gmm)
-				{
-					cat("Estimated by Generalized Method of Moments\n\n")
-				}
+				cat("Estimated by Generalized Method of Moments\n\n")
 			}
 			if (x$maxlike)
 			{
@@ -370,17 +367,14 @@ print.sienaFit <- function(x, tstat=TRUE, ...)
 				sem <- sem*sqrt(1 - (x$regrCor)^2)
 			}
 # mean.stats is exactly the same as x$estMeans
-			if (!is.null(x$gmm))
+			if (gmm(x))
 			{
-				if (x$gmm)
-				{
-					selection <- which(x$requestedEffects$type=='gmm' |
-									x$requestedEffects$type=='rate')
-				}
-				else
-				{
-					selection <- 1:nrow(x$requestedEffects)
-				}
+				selection <- which(x$requestedEffects$type=='gmm' |
+								x$requestedEffects$type=='rate')
+			}
+			else
+			{
+				selection <- 1:nrow(x$requestedEffects)
 			}
 			mymess1 <- paste(format(1:x$qq,width=3), '. ',
 					format(x$requestedEffects$functionName[selection], width = 56),
@@ -620,16 +614,13 @@ print.summary.sienaFit <- function(x, matrices=TRUE, ...)
 		covcor[lower.tri(covcor)] <- correl[lower.tri(correl)]
 		printMatrix(format(round(t(covcor),digits=3),width=12))
 		cat("\nDerivative matrix of expected statistics X by parameters:\n\n")
-		if (!is.null(x$gmm))
+		if (gmm(x))
 		{
-			if (x$gmm)
-			{
-				printMatrix(format(round(x$gamma,digits=3),width=12))
-			}
-			else
-			{
-				printMatrix(format(round(x$dfra,digits=3),width=12))
-			}
+			printMatrix(format(round(x$gamma,digits=3),width=12))
+		}
+		else
+		{
+			printMatrix(format(round(x$dfra,digits=3),width=12))
 		}
 		cat("\nCovariance matrix of X (correlations below diagonal):\n\n")
 		covcor <- x$msf
@@ -961,7 +952,7 @@ sienaFitThetaTable <- function(x, fromBayes=FALSE, tstat=FALSE, groupOnly=0, nfi
     {
         nrates <- 0
     }
-	if (!x$gmm)
+	if (!gmm(x))
 	{
 	  xp <- pp
 	  pp <- pp + nrates
@@ -1089,7 +1080,7 @@ sienaFitThetaTable <- function(x, fromBayes=FALSE, tstat=FALSE, groupOnly=0, nfi
     }
 	if (!is.null(x$covtheta))
 	{
-	  if (!x$gmm)
+	  if (!gmm(x))
 	  {
 	    ses <- sqrt(diag(x$covtheta))
 	    ses[x$fixed] <- NA
@@ -1106,7 +1097,7 @@ sienaFitThetaTable <- function(x, fromBayes=FALSE, tstat=FALSE, groupOnly=0, nfi
 		theta <- atl[[1]]
 		postSd <- sqrt(atl[[2]])
 	}
-  	if (x$gmm)
+  	if (gmm(x))
   	{
       theta <- x$theta[-which(x$requestedEffects$type=="gmm")]
 	}
@@ -1149,7 +1140,7 @@ sienaFitThetaTable <- function(x, fromBayes=FALSE, tstat=FALSE, groupOnly=0, nfi
         theEffects$effectName[theEffects$netType=='continuous'] <-
             contEffects$effectName
     }
-    if (!x$gmm)
+    if (!gmm(x))
       {
         mydf[nrates + (1:xp), 'row'] <-  1:xp
         mydf[nrates + (1:xp), 'type' ] <- ifelse(theEffects$type == "creation",
@@ -1157,7 +1148,7 @@ sienaFitThetaTable <- function(x, fromBayes=FALSE, tstat=FALSE, groupOnly=0, nfi
         mydf[nrates + (1:xp), 'text' ] <- theEffects$effectName
         mydf[nrates + (1:xp), 'value' ] <- theta
       }
-    else if (x$gmm)
+    else if (gmm(x))
     { 
       mydf[nrates + (1:xp), 'row'] <-  1:xp
       mydf[nrates + (1:xp), 'type' ] <- ifelse(theEffects$type == "creation",