+++ src/model/effects/EffectFactory.cpp
@@ -143,12 +143,12 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	Effect * pEffect = 0;
 	string effectName = pEffectInfo->effectName();
 
-	// Defined so we can later on differentiate between effects for 
+	// Defined so we can later on differentiate between effects for
 	// continuous and discrete dependent behavior variables
 	string variableName = pEffectInfo->variableName();
-	ContinuousLongitudinalData * pContinuousData = 
+	ContinuousLongitudinalData * pContinuousData =
         dynamic_cast<ContinuousLongitudinalData *>(this->lpData->pContinuousData(variableName));
-	
+
 	// Handle the user-defined interaction effects first.
 
 	if (pEffectInfo->pEffectInfo1())
@@ -208,17 +208,17 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	{
 		pEffect = new ReciprocityEffect(pEffectInfo);
 	}
-	else if (effectName == "newrecip") 
+	else if (effectName == "newrecip")
 	{
 		pEffect = new ReciprocityGMMEffect(pEffectInfo,
 				ReciprocityGMMEffect::NEW);
-	} 
-	else if (effectName == "realrecip") 
+	}
+	else if (effectName == "realrecip")
 	{
 		pEffect = new ReciprocityGMMEffect(pEffectInfo,
 				ReciprocityGMMEffect::REAL);
 	}
-	else if (effectName == "persistrecip") 
+	else if (effectName == "persistrecip")
 	{
 		pEffect = new ReciprocityGMMEffect(pEffectInfo,
 				ReciprocityGMMEffect::PERSISTENT);
@@ -975,7 +975,7 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	else if (effectName == "from")
 	{
 		pEffect = new GenericNetworkEffect(pEffectInfo,
-			new InStarFunction(pEffectInfo->interactionName1(), 
+			new InStarFunction(pEffectInfo->interactionName1(),
 							(pEffectInfo->internalEffectParameter() >= 2)));
 	}
 	else if (effectName == "fromMutual")
@@ -984,7 +984,7 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 			new ReciprocatedTwoPathFunction(pEffectInfo->interactionName1(),
 							(pEffectInfo->internalEffectParameter() >= 2)));
 	}
-	else if (effectName == "to") 
+	else if (effectName == "to")
 	{
 		pEffect = new GenericNetworkEffect(pEffectInfo,
 				new MixedTwoStepFunction(
@@ -1038,7 +1038,7 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 						FORWARD, BACKWARD, false));
 	}
 	else if (effectName == "mixedInXW")
-	{				
+	{
 		pEffect = new GenericNetworkEffect(pEffectInfo,
 				new MixedTwoStepFunction(
 						pEffectInfo->variableName(),
@@ -1407,22 +1407,50 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	{
 		pEffect = new QuadraticShapeEffect(pEffectInfo);
 	}
+	else if (effectName == "constant")
+	{
+		pEffect = new ConstantEffect(pEffectInfo);
+	}
 	else if ((effectName == "threshold") | (effectName == "threshold2") |
-	         (effectName == "threshold3") |(effectName == "threshold3"))
+	         (effectName == "threshold3") |(effectName == "threshold4"))
 	{
 		pEffect = new ThresholdShapeEffect(pEffectInfo);
 	}
 	else if (effectName == "avSim")
 	{
-		pEffect = new SimilarityEffect(pEffectInfo, true, false, false);
+		pEffect = new SimilarityEffect(pEffectInfo, true, false, false, true, true);
+	}
+	else if (effectName == "avAttHigher")
+	{
+		pEffect = new SimilarityEffect(pEffectInfo, true, false, false, true, false);
+	}
+	else if (effectName == "avAttLower")
+	{
+		pEffect = new SimilarityEffect(pEffectInfo, true, false, false, false, true);
 	}
 	else if (effectName == "avSim_gmm")
 	{
-		pEffect = new SimilarityEffect(pEffectInfo, true, false, false,true);
+		pEffect = new SimilarityEffect(pEffectInfo, true, false, false, true, true, true);
 	}
 	else if (effectName == "totSim")
 	{
-		pEffect = new SimilarityEffect(pEffectInfo, false, false, false);
+		pEffect = new SimilarityEffect(pEffectInfo, false, false, false, true, true);
+	}
+	else if (effectName == "totAttHigher")
+	{
+		pEffect = new SimilarityEffect(pEffectInfo, false, false, false, true, false);
+	}
+	else if (effectName == "totAttLower")
+	{
+		pEffect = new SimilarityEffect(pEffectInfo, false, false, false, false, true);
+	}
+	else if (effectName == "avInSim")
+	{
+		pEffect = new SimilarityIndegreeEffect(pEffectInfo, true, false);
+	}
+	else if (effectName == "totInSim")
+	{
+		pEffect = new SimilarityIndegreeEffect(pEffectInfo, false, false);
 	}
 	else if (effectName == "indeg")
 	{
@@ -1433,7 +1461,7 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 		else
 		{
 		pEffect = new IndegreeEffect(pEffectInfo);
-	}
+		}
 	}
 	else if (effectName == "indegSqrt")
 	{
@@ -1451,7 +1479,7 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 		else
 		{
 		pEffect = new OutdegreeEffect(pEffectInfo);
-	}
+		}
 	}
 	else if (effectName == "outdegSqrt")
 	{
@@ -1543,7 +1571,7 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	else if (effectName == "settingLogCreationActD")
 	{
 		pEffect = new SettingSizeEffect(pEffectInfo, true, true, false, true, true);
-	}	
+	}
 	else if (effectName == "avSimRecip")
 	{
 		pEffect = new ReciprocatedSimilarityEffect(pEffectInfo, true, false);
@@ -1554,11 +1582,19 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	}
 	else if (effectName == "avSimPopAlt")
 	{
-		pEffect = new SimilarityEffect(pEffectInfo, true, true, false);
+		pEffect = new SimilarityEffect(pEffectInfo, true, true, false, true, true);
 	}
 	else if (effectName == "totSimPopAlt")
 	{
-		pEffect = new SimilarityEffect(pEffectInfo, false, true, false);
+		pEffect = new SimilarityEffect(pEffectInfo, false, true, false, true, true);
+	}
+	else if (effectName == "avInSimPopAlt")
+	{
+		pEffect = new SimilarityIndegreeEffect(pEffectInfo, true, true);
+	}
+	else if (effectName == "totInSimPopAlt")
+	{
+		pEffect = new SimilarityIndegreeEffect(pEffectInfo, false, true);
 	}
 	else if (effectName == "popAlt")
 	{
@@ -1579,7 +1615,7 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 			pEffect = new AverageAlterContinuousEffect(pEffectInfo);
 		}
 		else
-		{	
+		{
 		pEffect = new AverageAlterEffect(pEffectInfo, true, false);
 	}
 	}
@@ -1734,11 +1770,11 @@ Effect * EffectFactory::createEffect(const EffectInfo * pEffectInfo) const
 	}
 //	else if (effectName == "RRDeg")
 //	{
-//		pEffect = new DoubleRecDegreeBehaviorEffect(pEffectInfo, 2); // crashes
+//		pEffect = new DoubleRecDegreeBehaviorEffect(pEffectInfo, 2); // leads to error
 //	}
 	else if (effectName == "avSimPopEgo")
 	{
-		pEffect = new SimilarityEffect(pEffectInfo, true, false, true);
+		pEffect = new SimilarityEffect(pEffectInfo, true, false, true, true, true);
 	}
 	else if (effectName == "effFrom")
 	{